Changed CategoryRepository, ItemCategoryRepository, OrderRepository and ItemRepository interfaces to use Long as ID.

+++ src/main/java/com/answerdigital/benhession/academy/answerkingweek2/repositories/CategoryRepository.java
@@ -4,19 +4,16 @@ import com.answerdigital.benhession.academy.answerkingweek2.model.Category;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.HashSet;
-import java.util.Optional;
 import java.util.Set;
 
 @Repository
-public interface CategoryRepository extends CrudRepository<Category, Integer> {
-
+public interface CategoryRepository extends CrudRepository<Category, Long>
+{
     boolean existsByName(String name);
 
-    boolean existsByNameAndIdIsNot(String name, Integer id);
+    boolean existsByNameAndIdIsNot(String name, Long id);
 
     Set<Category> findAll();
 
-    int countByIdIn(Set<Integer> ids);
-
+    int countByIdIn(Set<Long> ids);
 }
